# Speed up Arr::unique